Rewrite optimization barrier to use new asm! macro
Since llvm_asm! is deprecated, I rewrote the optimization barrier to use
the new asm! macro. Hopefully this macro will be stabilized in the near
future.

I also extracted the optimization barrier to a separate module and
documented it as a utility function. This had the added bonus of making
explicit the transmute from `u8` to `bool` in `TpBool::expose`.